Add spider to scrap product details
