+++ closetcoach_crawler/scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = myntra_spider.settings
+default = closetcoach_crawler.myntra_spider.settings
 
 [deploy]
 #url = http://localhost:6800/
-project = myntra_spider
+project = closetcoach_crawler
