--- data/spiders/myntra_spider/myntra_spider/run_product_page_url_scraper_spider.sh
@@ -1,128 +1,0 @@
-cd $1
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-tshirts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-casual-shirts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-formal-shirts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sweatshirts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sweaters
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-jackets
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-blazers
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-suits
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/rain-jacket
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-jeans
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-casual-trousers
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-formal-trousers
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/mens-shorts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-trackpants
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-casual-shoes
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sports-shoes
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-formal-shoes
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sneakers
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sandals
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-flip-flops
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-socks
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sports-shoes
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sports-sandals
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sports-wear-tshirts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-trackpants-shorts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-tracksuits
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sports-jackets
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sports-accessories
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-swimwear
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-kurtas
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/sherwani
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/nehru-jackets
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/dhoti
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/men-sunglasses
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-kurtas-kurtis-suits
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/ethnic-tops
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/saree
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-ethnic-wear
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-ethnic-bottomwear
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/skirts-palazzos
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/dress-material
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/lehenga-choli
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/dupatta-shawl
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-jackets
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/dresses
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/tops
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/myntra-fashion-store
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-jeans
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-trousers
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-shorts-skirts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/myntra-fashion-store
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/playsuit
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/jumpsuits
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-shrugs
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-sweaters-sweatshirts
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-jackets-coats
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-blazers-waistcoats
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/flats
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-casual-shoes
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-heels
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-boots-menu
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-sports-shoes
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-sportswear-clothing
-
-scrapy crawl product_page_url_scraper -a start_urls=https://www.myntra.com/women-sports-shoes
